+++ src/components/statistic/index.ts
@@ -1,0 +1,2 @@
+export { MPStatistic } from './MPStatistic';
+export type { MPStatisticDeltaFormat, MPStatisticProps, MPStatisticSlot } from './MPStatistic';